navbar styled with links mobile desktop
