+++ src/components/NavBar/NavBar.jsx
@@ -1,26 +1,16 @@
 import React, { useState } from "react";
-import { useNavigate, Link } from "react-router-dom";
+import { Link } from "react-router-dom";
+import NavBarNotLogged from "./NavBarNotLogged";
+import NavBarLogged from "./NavBarLogged";
 import SearchModalMobile from "../SearchNav/SearchModalMobile";
 import "./NavBar.scss";
 import logo from "../../images/what toDo.png";
-import {
-  FaSearch,
-  FaHeart,
-  FaRegHeart,
-  FaSignInAlt,
-  FaSignOutAlt,
-  FaRegEdit,
-} from "react-icons/fa";
+import { FaSearch } from "react-icons/fa";
 
 export default function NavBar() {
   const [modal, setModal] = useState(false);
   const Toggle = () => setModal(!modal); //to open mobile search modal
-  const isLogin = localStorage.clear();
-  const navigate = useNavigate();
-  const logoutHandler = () => {
-    localStorage.clear();
-    navigate("/");
-  };
+  const isLogin = localStorage.getItem("isLogin");
 
   return (
     <div className="navbar">
@@ -29,31 +19,8 @@ export default function NavBar() {
           <img src={logo} alt="Logo WhatToDo" />
         </Link>
       </div>
-      <div className="nav">
-        {isLogin ? (
-          <Link to={"/create-listing"}>
-            <FaRegEdit className="create-event-mobile-icon" />
-          </Link>
-        ) : null}
 
-        {isLogin ? (
-          <button className="btn-navbar">
-            <Link to="/create-listing" className="link-navbar">
-              LIST AN EVENT
-            </Link>
-          </button>
-        ) : null}
-        {isLogin ? (
-          <button className="btn-navbar" onClick={logoutHandler}>
-            LOG OUT!
-          </button>
-        ) : (
-          <button className="btn-navbar">
-            <Link to="/login" className="link-navbar">
-              LOG IN!
-            </Link>
-          </button>
-        )}
+      <div className="nav-wrapper">
         <button className="search-mobile" onClick={() => Toggle()}>
           <FaSearch className="mobile-search-icon" />
         </button>
@@ -62,18 +29,11 @@ export default function NavBar() {
           title="WHAT TO DO TODAY...?"
           close={Toggle}
         />
-        <FaRegHeart className="heart-icon" />
 
-        {isLogin ? (
-          <Link to={"/event"}>
-            <FaSignOutAlt className="login-icon" />
-          </Link>
-        ) : (
-          <Link to={"/login"}>
-            <FaSignInAlt className="login-icon" />
-          </Link>
-        )}
+        {isLogin ? <NavBarLogged /> : <NavBarNotLogged />}
+
       </div>
+      
     </div>
   );
 }
