+++ src/components/NavBar/NavBar.scss
@@ -1,5 +1,3 @@
-@import "../../scss/variable.scss";
-
 @media (max-width: 820px) {
 .navbar {
     height: 80px;
@@ -19,9 +17,8 @@
         }
     }
 
-    .nav {
-        width: 45%;
-        margin-right: 20px;
+    .nav-wrapper {
+        width: 54%;
         display: flex;
         align-items: center;
         justify-content: space-between;
@@ -45,7 +42,7 @@
         border: none;
         transition: all 0.25s ease;
       }
-  }
+   }
 
 }
 
@@ -68,8 +65,7 @@
             }
         }
         
-        .nav {
-            width: 40%;
+        .nav-wrapper {
             display: flex;
             align-items: center;
             justify-content: space-between;
@@ -83,34 +79,7 @@
             .heart-icon {
                 font-size: 3em;
             }
-            .btn-navbar{
-                display: unset;
-                background-color: $color-white;
-                border-radius: 15px;
-                width: 120px;
-                height: 30px;
-                &:hover {
-                    color: #fff!important;
-                    border-color: #000;
-                    background-color: #000;
-                    transition: color .25s ease-in-out;
-                    transition: background-color .25s ease-in-out;
-                }
-            }
-
-            .link-navbar{
-            text-decoration: none;
-            color: black;
-            background-color: transparent;
-            padding-top: 10px;
-            padding-bottom: 10px;
-            &:hover {
-                color: #fff!important;
-                transition: color .25s ease-in-out;
-                transition: background-color .25s ease-in-out;
-            }
-          }
-     
+       
         }
     }
 }
