+++ src/components/NavBar/NavBarLogged.scss
@@ -1,0 +1,52 @@
+@import "../../scss/variable.scss";
+
+@media (max-width: 820px) {
+  .nav {
+    width: 200px;
+    margin-right: 20px;
+    margin-left: 0;
+    display: flex;
+    align-items: center;
+    justify-content: space-around;
+  }
+}
+
+@media (min-width: 1281px) {
+  .nav {
+    width: 100%;
+    .btn-navbar {
+      display: unset;
+      background-color: $color-white;
+      border-radius: 15px;
+      width: 150px;
+      height: 30px;
+      margin-right: 20px;
+      &:hover {
+        color: #fff !important;
+        border-color: #000;
+        background-color: #000;
+        transition: color 0.25s ease-in-out;
+        transition: background-color 0.25s ease-in-out;
+      }
+    }
+    .link-navbar {
+      width: 150px;
+      text-decoration: none;
+      color: black;
+      background-color: transparent;
+      padding-top: 10px;
+      padding-bottom: 10px;
+      &:hover {
+        color: #fff !important;
+        transition: color 0.25s ease-in-out;
+        transition: background-color 0.25s ease-in-out;
+      }
+    }
+    .heart-icon {
+      font-size: 0.8em;
+      text-decoration: none;
+      color: black;
+      margin-right: 0;
+    }
+  }
+}
